+++ src/VTKViewport/vtkInteractorStyle3DCrosshairs.js
@@ -1,200 +1,207 @@
 import macro from 'vtk.js/Sources/macro';
 import vtkInteractorStyleTrackballCamera from 'vtk.js/Sources/Interaction/Style/InteractorStyleTrackballCamera';
+import Constants from 'vtk.js/Sources/Rendering/Core/InteractorStyle/Constants';
+import vtkPlaneManipulator from 'vtk.js/Sources/Widgets/Manipulators/PlaneManipulator';
 import vtkCoordinate from 'vtk.js/Sources/Rendering/Core/Coordinate';
-import vtkPointPicker from 'vtk.js/Sources/Rendering/Core/PointPicker';
-import vtkImageMarchingSquares from 'vtk.js/Sources/Filters/General/ImageMarchingCubes';
-import vtkImageMapper from 'vtk.js/Sources/Rendering/Core/ImageMapper';
-import vtkActor from 'vtk.js/Sources/Rendering/Core/Actor';
+import vtkBoundingBox from 'vtk.js/Sources/Common/DataModel/BoundingBox';
+import { VTKAxis } from '@vtk-viewport';
 
-import VTKAxis from './VTKAxis';
+const { States } = Constants;
 
-function vtkInteractorStyle3DCrosshairs(publicAPI, model) {
-  //Set classname
+// ----------------------------------------------------------------------------
+// Global methods
+// ----------------------------------------------------------------------------
+
+// ----------------------------------------------------------------------------
+// vtkInteractorStyleCrosshairsMarchingCubes methods
+// ----------------------------------------------------------------------------
+
+function vtkInteractorStyleCrosshairsMarchingCubes(publicAPI, model) {
+  // Set our className
   model.classHierarchy.push('vtkInteractorStyle3DCrosshairs');
+  model.planeManipulator = vtkPlaneManipulator.newInstance();
 
-  //get copy of default publicAPI
-  const superAPI = Object.assign({}, publicAPI);
+  publicAPI.resetCrosshairs = () => {
+    const { apis, apiIndex } = model;
+    const api = apis[apiIndex];
 
-  const crosshairs = new VTKAxis(0, 0, 0, 0);
-  publicAPI.setCrosshairs(crosshairs);
+    api.svgWidgets.crosshairsWidget.moveCrosshairs(
+      publicAPI.getCenter(),
+      apis,
+      apiIndex
+    );
 
-  publicAPI.moveCrosshairs = callData => {
-    const api = model.apis[model.apiIndex];
-
-    const mousePos = callData.position;
-
-    const coords = vtkCoordinate.newInstance();
-    coords.setCoordinateSystemToDisplay();
-    coords.setValue(mousePos.x, mousePos.y, mousePos.z);
-
-    const actors = publicAPI.getActors();
-    const renderer = api.genericRenderWindow.getRenderer();
-    const screenPoint = [mousePos.x, mousePos.y, 0];
-
-    const picker = vtkPointPicker.newInstance();
-    picker.setPickFromList(false);
-    // picker.initializePickList();
-    // actors.forEach((actor, i) => {
-    //   picker.addPickList(actor, renderer);
-    // });
-
-    const picked = picker.pick(screenPoint, renderer);
-    const pointId = picker.getPointId();
-
-    const worldPos = picker
-      .getActors()[0]
-      .getMapper()
-      .getInputData()
-      .getPoints()
-      .getPoint(pointId);
-
-    model.apis.forEach((api, i) => {
-      const istyle = api.genericRenderWindow
-        .getRenderWindow()
-        .getInteractor()
-        .getInteractorStyle();
-
-      istyle.updateCrosshairs(worldPos);
-    });
+    publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
   };
 
-  publicAPI.updateCrosshairs = worldPos => {
-    publicAPI.getCrosshairs().setPoint(...worldPos);
-    publicAPI
-      .getApis()
-      [publicAPI.getApiIndex()].genericRenderWindow.getRenderWindow()
-      .render();
+  publicAPI.moveCrosshairs = callData => {
+    const { apis, apiIndex } = model;
+    const api = apis[apiIndex];
+
+    const pos = callData.position;
+    const renderer = callData.pokedRenderer;
+
+    const dPos = vtkCoordinate.newInstance();
+    dPos.setCoordinateSystemToDisplay();
+
+    dPos.setValue(pos.x, pos.y, 0);
+    const worldCoords = dPos.getComputedWorldValue(renderer);
+
+    api.svgWidgets.crosshairsWidget.moveCrosshairs(worldCoords, apis, apiIndex);
+
+    publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
+  };
+
+  const superHandleMouseMove = publicAPI.handleMouseMove;
+  publicAPI.handleMouseMove = callData => {
+    if (model.state === States.IS_WINDOW_LEVEL) {
+      publicAPI.moveCrosshairs(callData);
+    } else if (superHandleMouseMove) {
+      superHandleMouseMove(callData);
+    }
+  };
+
+  publicAPI.superHandleLeftButtonPress = publicAPI.handleLeftButtonPress;
+  publicAPI.handleLeftButtonPress = callData => {
+    if (!callData.shiftKey && !callData.controlKey) {
+      if (model.actor) {
+        publicAPI.moveCrosshairs(callData);
+        publicAPI.startWindowLevel();
+      }
+    } else if (publicAPI.superHandleLeftButtonPress) {
+      publicAPI.superHandleLeftButtonPress(callData);
+    }
+  };
+
+  publicAPI.handleRightButtonPress = callData => {
+    publicAPI.superHandleLeftButtonPress(callData);
+  };
+
+  publicAPI.superHandleLeftButtonRelease = publicAPI.handleLeftButtonRelease;
+  publicAPI.handleRightButtonRelease = callData => {
+    publicAPI.superHandleLeftButtonRelease(callData);
+  };
+
+  publicAPI.superHandleLeftButtonRelease = publicAPI.handleLeftButtonRelease;
+  publicAPI.handleLeftButtonRelease = () => {
+    switch (model.state) {
+      case States.IS_WINDOW_LEVEL:
+        publicAPI.endWindowLevel();
+        break;
+
+      default:
+        publicAPI.superHandleLeftButtonRelease();
+        break;
+    }
+  };
+
+  // publicAPI.setCrosshairWidth = (width) => {
+  //   if (publicAPI.getAxis()){
+  //     publicAPI.getAxis().setWidth(width)
+  //   }
+  //   else{
+  //     publicAPI.setCrosshairWidth(width)
+  //   }
+  // }
+
+  // Slice normal is just camera DOP
+  publicAPI.getSliceNormal = () => {
+    if (model.actor && model.interactor) {
+      const renderer = model.interactor.getCurrentRenderer();
+      const camera = renderer.getActiveCamera();
+      return camera.getDirectionOfProjection();
+    }
+    return [0, 0, 0];
   };
 
   publicAPI.setActor = actor => {
-    actor.getProperty().setOpacity(0.5);
-    superAPI.setActor(actor);
+    model.actor = actor;
+    model.axis = new VTKAxis(0, 0, 0, publicAPI.getCrosshairWidth());
+
+    publicAPI.set('axis', model.axis);
 
     const renderer = model.interactor.getCurrentRenderer();
-
-    const bounds = actor.getBounds();
-    const width =
-      Math.max(
-        Math.abs(bounds[1] - bounds[0]),
-        Math.abs(bounds[3] - bounds[2]),
-        Math.abs(bounds[5] - bounds[4])
-      ) * 2;
-
-    const pos = actor.getCenter();
-    const crosshairs = publicAPI.getCrosshairs();
-    crosshairs.actors.forEach(renderer.addActor);
-    crosshairs.setPoint(...pos);
-    crosshairs.setWidth(width);
-  };
-
-  publicAPI.setActors = actors => {
-    actors.forEach((actor, i) => {
-      actor.getProperty().setOpacity(0.5);
-    });
-    superAPI.setActors(actors);
-  };
-
-  publicAPI.addCrosshairSlices = apis => {
-    const renderer = model.interactor.getCurrentRenderer();
-
-    apis.forEach((api, i) => {
-      if (api.type == 'VIEW2D') {
-        api.actors.forEach(renderer.addActor);
-      }
-    });
-  };
-
-  publicAPI.handleLeftButtonPress = callData => {
-    if (callData.shiftKey || callData.altKey || callData.controlKey) {
-      //If not just pressing the mouse button, do whatever trackball camera would normally do
-      superAPI.handleLeftButtonPress(callData);
+    const camera = renderer.getActiveCamera();
+    if (actor) {
+      // prevent zoom manipulator from messing with our focal point
+      camera.setFreezeFocalPoint(true);
     } else {
-      //otherwise, move the crosshairs
-      publicAPI.moveCrosshairs(callData);
-      model.movingCrosshairs = true;
+      camera.setFreezeFocalPoint(false);
     }
+
+    console.log({ model, publicAPI });
+
+    publicAPI.get('axis').axis.actors.map(actor => renderer.addActor(actor));
   };
 
-  publicAPI.handleLeftButtonRelease = callData => {
-    model.movingCrosshairs = false;
+  publicAPI.setImageActor = actor => {
+    model.actor = actor;
+    model.axis = new VTKAxis(0, 0, 0, publicAPI.getCrosshairWidth());
 
-    superAPI.handleLeftButtonRelease(callData);
-  };
+    publicAPI.set('axis', model.axis);
 
-  publicAPI.handleMouseMove = callData => {
-    if (
-      !(callData.shiftKey || callData.altKey || callData.controlKey) &&
-      model.movingCrosshairs
-    ) {
-      //otherwise, move the crosshairs
-      publicAPI.moveCrosshairs(callData);
-    } else {
-      //If not just pressing the mouse button, do whatever trackball camera would normally do
-      superAPI.handleMouseMove(callData);
-    }
-  };
-
-  publicAPI.toggleCrosshairs = () => {
-    const crosshairs = publicAPI.getCrosshairs();
     const renderer = model.interactor.getCurrentRenderer();
 
-    crosshairs.actors.forEach((actor, i) => {
-      if (renderer.getActors().includes(actor)) {
-        renderer.removeActor(actor);
-      } else {
-        renderer.addActor(actor);
-      }
-    });
-
-    renderer.getRenderWindow().render();
+    publicAPI.get('axis').axis.actors.map(actor => renderer.addActor(actor));
   };
 
-  publicAPI.toggleCrosshairSlices = () => {
+  publicAPI.getImageActor = () => {
+    return null;
+  };
+
+  publicAPI.getSliceCenter = () => {
+    //set to center of current slice
+    return vtkBoundingBox.getCenter(model.actor.getBoundsForSlice());
+  };
+
+  publicAPI.getCemter = () => {
+    // Get viewport and get its center.
     const renderer = model.interactor.getCurrentRenderer();
+    const view = renderer.getRenderWindow().getViews()[0];
+    const dims = view.getViewportSize(renderer);
+    const dPos = vtkCoordinate.newInstance();
 
-    model.apis.forEach((api, i) => {
-      if (api.type == 'VIEW2D') {
-        const actor = api.actors[0];
-        if (renderer.getActors().includes(actor)) {
-          renderer.removeActor(actor);
-        } else {
-          renderer.addActor(actor);
-        }
-      }
-    });
+    dPos.setCoordinateSystemToDisplay();
 
-    renderer.getRenderWindow().render();
+    dPos.setValue(0.5 * dims[0], 0.5 * dims[1], 0);
+    return dPos.getComputedWorldValue(renderer);
   };
-
-  publicAPI.handleRightButtonPress = superAPI.handleLeftButtonPress;
-  publicAPI.handleRightButtonRelease = superAPI.handleLeftButtonRelease;
 }
 
-//Class defaults
+// ----------------------------------------------------------------------------
+// Object factory
+// ----------------------------------------------------------------------------
+
 const DEFAULT_VALUES = {};
 
+// ----------------------------------------------------------------------------
+
 export function extend(publicAPI, model, initialValues = {}) {
-  //combine all default values, initial values, and model into one object
   Object.assign(model, DEFAULT_VALUES, initialValues);
 
+  // Inheritance
   vtkInteractorStyleTrackballCamera.extend(publicAPI, model, initialValues);
 
-  //assign variables that can be set and gotten
   macro.setGet(publicAPI, model, [
+    'callback',
     'apis',
     'apiIndex',
-    'crosshairs',
     'actor',
-    'actors',
+    'axis',
+    'crosshairWidth',
   ]);
 
-  //add function
-  vtkInteractorStyle3DCrosshairs(publicAPI, model);
+  // Object specific methods
+  vtkInteractorStyleCrosshairsMarchingCubes(publicAPI, model);
 }
+
+// ----------------------------------------------------------------------------
 
 export const newInstance = macro.newInstance(
   extend,
-  'vtkInteractorStyle3DCrosshairs'
+  'vtkInteractorStyleCrosshairsMarchingCubes'
 );
 
-export default { newInstance, extend };
+// ----------------------------------------------------------------------------
+
+export default Object.assign({ newInstance, extend });
